% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym_test.R
\docType{data}
\name{asym_test}
\alias{asym_test}
\title{Asymmetry Parameters of \code{nnadic} Test Data}
\format{
\subsection{\code{asym_test}}{

A matrix with 4,000 rows and 2 columns:
\describe{
\item{asym_param1}{First asymmetry parameter}
\item{asym_param2}{Second asymmetry paramter}
...
}
}
}
\source{
data_generation_four_types.R
}
\usage{
asym_test
}
\description{
The only model with asymmetry is the asymmetric logistic model. All others
have asymmetry parameters (0,0)
}
\keyword{datasets}
