% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_chiBar_four.R
\docType{data}
\name{chi_chiBar_four}
\alias{chi_chiBar_four}
\title{chi or chiBar values of \code{nnadic} Test Data}
\format{
\subsection{\code{chi_chiBar_four}}{

A vector with 4000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
chi_chiBar_four
}
\description{
For each data set in the test data the dependence measure can be found here.
If the data are asymptotically dependent then chi is reported. Otherwise
chiBar is reported.
}
\keyword{datasets}
