% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_four.R
\docType{data}
\name{chi_four}
\alias{chi_four}
\title{chi values of \code{nnadic} Test Data}
\format{
\subsection{\code{chi_four}}{

A vector with 4000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
chi_four
}
\description{
chi values for each data set in the test data. Zeros indicate asymptotic
independence
}
\keyword{datasets}
