% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiBar_four.R
\docType{data}
\name{chiBar_four}
\alias{chiBar_four}
\title{chiBar values of \code{nnadic} Test Data}
\format{
\subsection{\code{chiBar_four}}{

A vector with 4000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
chiBar_four
}
\description{
chiBar values for each data set in the test data. Ones indicate the data
were from the logistic model (asymptotically dependent). Zeros
(incorrectly) indicate asymmetric logistic model.
}
\keyword{datasets}
