% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_training_data.R
\name{combine_training_data}
\alias{combine_training_data}
\title{Combine the eight different training datasets into one}
\usage{
combine_training_data()
}
\value{
puts \code{train_data_four} into the environment
}
\description{
The datasets were split for faster loading but are useful combined
}
\examples{
combine_training_data()
}
