% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_four.R
\docType{data}
\name{test_data_four}
\alias{test_data_four}
\title{Testing data used in evaluation of \code{nnadic} model}
\format{
\subsection{\code{test_data_four}}{

An array with 4000 data sets of 500 bivariate points
}
}
\source{
data_generation_four_types.R
}
\usage{
test_data_four
}
\description{
Testing data used in evaluation of \code{nnadic} model
}
\keyword{datasets}
