% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dep_four.R
\docType{data}
\name{test_dep_four}
\alias{test_dep_four}
\title{Dependence parameter values used in \code{nnadic} Test Data generation}
\format{
\subsection{\code{test_dep_four}}{

A vector with 4000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
test_dep_four
}
\description{
Dependence parameter values used in \code{nnadic} Test Data generation
}
\keyword{datasets}
