% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_response_four.R
\docType{data}
\name{test_response_four}
\alias{test_response_four}
\title{Response values for \code{nnadic} Test Data}
\format{
\subsection{\code{test_response_four}}{

A vector with 4000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
test_response_four
}
\description{
Zeros indicate asymptotic dependence, ones indicate asymptotic independence
}
\keyword{datasets}
