% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_data_four3.R
\docType{data}
\name{train_data_four3}
\alias{train_data_four3}
\title{Training data used to train \code{nnadic} model}
\format{
\subsection{\code{train_data_four3}}{

The third of four arrays with 8000 data sets of 500 bivariate points
}
}
\source{
data_generation_four_types.R
}
\usage{
train_data_four3
}
\description{
Training data used to train \code{nnadic} model
}
\keyword{datasets}
