% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_response_four.R
\docType{data}
\name{train_response_four}
\alias{train_response_four}
\title{Response values for \code{nnadic} training data}
\format{
\subsection{\code{train_response_four}}{

A vector with 32000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
train_response_four
}
\description{
Zeros indicate asymptotic dependence, ones indicate asymptotic independence
}
\keyword{datasets}
