% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_data_four.R
\docType{data}
\name{valid_data_four}
\alias{valid_data_four}
\title{Validation data used in training of \code{nnadic} model}
\format{
\subsection{\code{valid_data_four}}{

An array with 4000 data sets of 500 bivariate points
}
}
\source{
data_generation_four_types.R
}
\usage{
valid_data_four
}
\description{
Validation data used in training of \code{nnadic} model
}
\keyword{datasets}
