% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_response_four.R
\docType{data}
\name{valid_response_four}
\alias{valid_response_four}
\title{Response values for \code{nnadic} validation data}
\format{
\subsection{\code{valid_response_four}}{

A vector with 32000 entries
}
}
\source{
data_generation_four_types.R
}
\usage{
valid_response_four
}
\description{
Zeros indicate asymptotic dependence, ones indicate asymptotic independence
}
\keyword{datasets}
